Switch from password and salt to an encryption key
